initial file split
