+++ wandb/sdk/interface/data_types/audio.py
@@ -1,0 +1,96 @@
+class Audio(BatchableMedia):
+    """
+        Wandb class for audio clips.
+
+        Arguments:
+            data_or_path (string or numpy array): A path to an audio file
+                or a numpy array of audio data.
+            sample_rate (int): Sample rate, required when passing in raw
+                numpy array of audio data.
+            caption (string): Caption to display with audio.
+    """
+
+    def __init__(self, data_or_path, sample_rate=None, caption=None):
+        """Accepts a path to an audio file or a numpy array of audio data.
+        """
+        super(Audio, self).__init__()
+        self._duration = None
+        self._sample_rate = sample_rate
+        self._caption = caption
+
+        if isinstance(data_or_path, six.string_types):
+            self._set_file(data_or_path, is_tmp=False)
+        else:
+            if sample_rate == None:
+                raise ValueError(
+                    'Argument "sample_rate" is required when instantiating wandb.Audio with raw data.'
+                )
+
+            soundfile = util.get_module(
+                "soundfile",
+                required='Raw audio requires the soundfile package. To get it, run "pip install soundfile"',
+            )
+
+            tmp_path = os.path.join(MEDIA_TMP.name, util.generate_id() + ".wav")
+            soundfile.write(tmp_path, data_or_path, sample_rate)
+            self._duration = len(data_or_path) / float(sample_rate)
+
+            self._set_file(tmp_path, is_tmp=True)
+
+    @classmethod
+    def get_media_subdir(cls):
+        return os.path.join("media", "audio")
+
+    def to_json(self, run):
+        json_dict = super(Audio, self).to_json(run)
+        json_dict.update(
+            {
+                "_type": "audio-file",
+                "sample_rate": self._sample_rate,
+                "caption": self._caption,
+            }
+        )
+        return json_dict
+
+    @classmethod
+    def seq_to_json(cls, seq, run, key, step):
+        audio_list = list(seq)
+
+        sf = util.get_module(
+            "soundfile",
+            required="wandb.Audio requires the soundfile package. To get it, run: pip install soundfile",
+        )
+        base_path = os.path.join(run.dir, "media", "audio")
+        util.mkdir_exists_ok(base_path)
+        meta = {
+            "_type": "audio",
+            "count": len(audio_list),
+            "audio": [a.to_json(run) for a in audio_list],
+        }
+        sample_rates = cls.sample_rates(audio_list)
+        if sample_rates:
+            meta["sampleRates"] = sample_rates
+        durations = cls.durations(audio_list)
+        if durations:
+            meta["durations"] = durations
+        captions = cls.captions(audio_list)
+        if captions:
+            meta["captions"] = captions
+
+        return meta
+
+    @classmethod
+    def durations(cls, audio_list):
+        return [a._duration for a in audio_list]
+
+    @classmethod
+    def sample_rates(cls, audio_list):
+        return [a._sample_rate for a in audio_list]
+
+    @classmethod
+    def captions(cls, audio_list):
+        captions = [a._caption for a in audio_list]
+        if all(c is None for c in captions):
+            return False
+        else:
+            return ["" if c == None else c for c in captions]
