+++ wandb/sdk/interface/data_types/image_mask.py
@@ -1,0 +1,115 @@
+class ImageMask(Media):
+    """
+    Wandb class for image masks, useful for segmentation tasks
+    """
+
+    artifact_type = "mask"
+
+    def __init__(self, val, key, **kwargs):
+        """
+        Args:
+            val (dict): dictionary following 1 of two forms: 
+            {
+                "mask_data": 2d array of integers corresponding to classes,
+                "class_labels": optional mapping from class ids to strings {id: str}
+            }
+
+            {
+                "path": path to an image file containing integers corresponding to classes,
+                "class_labels": optional mapping from class ids to strings {id: str}
+            }
+            key (str): id for set of masks
+        """
+        super(ImageMask, self).__init__()
+
+        if "path" in val:
+            self._set_file(val["path"])
+        else:
+            np = util.get_module(
+                "numpy", required="Semantic Segmentation mask support requires numpy"
+            )
+            # Add default class mapping
+            if not "class_labels" in val:
+                classes = np.unique(val["mask_data"]).astype(np.int32).tolist()
+                class_labels = dict((c, "class_" + str(c)) for c in classes)
+                val["class_labels"] = class_labels
+
+            self.validate(val)
+            self._val = val
+            self._key = key
+
+            ext = "." + self.type_name() + ".png"
+            tmp_path = os.path.join(MEDIA_TMP.name, util.generate_id() + ext)
+
+            PILImage = util.get_module(
+                "PIL.Image",
+                required='wandb.Image needs the PIL package. To get it, run "pip install pillow".',
+            )
+            image = PILImage.fromarray(val["mask_data"].astype(np.int8), mode="L")
+
+            image.save(tmp_path, transparency=None)
+            self._set_file(tmp_path, is_tmp=True, extension=ext)
+
+    def bind_to_run(self, run, key, step, id_=None):
+        # bind_to_run key argument is the Image parent key
+        # the self._key value is the mask's sub key
+        super(ImageMask, self).bind_to_run(run, key, step, id_=id_)
+        class_labels = self._val["class_labels"]
+
+        run._add_singleton(
+            "mask/class_labels", key + "_wandb_delimeter_" + self._key, class_labels
+        )
+
+    def get_media_subdir(self):
+        return os.path.join("media", "images", self.type_name())
+
+    @classmethod
+    def from_json(cls, json_obj, source_artifact):
+        return cls(
+            {"path": source_artifact.get_path(json_obj["path"]).download()}, key="",
+        )
+
+    def to_json(self, run_or_artifact):
+        json_dict = super(ImageMask, self).to_json(run_or_artifact)
+        wandb_run, wandb_artifacts = _safe_sdk_import()
+
+        if isinstance(run_or_artifact, wandb_run.Run):
+            json_dict["_type"] = self.type_name()
+            return json_dict
+        elif isinstance(run_or_artifact, wandb_artifacts.Artifact):
+            # Nothing special to add (used to add "digest", but no longer used.)
+            return json_dict
+        else:
+            raise ValueError("to_json accepts wandb_run.Run or wandb_artifact.Artifact")
+
+    def type_name(self):
+        return "mask"
+
+    def validate(self, mask):
+        np = util.get_module(
+            "numpy", required="Semantic Segmentation mask support requires numpy"
+        )
+        # 2D Make this work with all tensor(like) types
+        if not "mask_data" in mask:
+            raise TypeError(
+                'Missing key "mask_data": A mask requires mask data(A 2D array representing the predctions)'
+            )
+        else:
+            error_str = "mask_data must be a 2d array"
+            shape = mask["mask_data"].shape
+            if len(shape) != 2:
+                raise TypeError(error_str)
+            if not (
+                (mask["mask_data"] >= 0).all() and (mask["mask_data"] <= 255).all()
+            ) and issubclass(mask["mask_data"].dtype.type, np.integer):
+                raise TypeError("Mask data must be integers between 0 and 255")
+
+        # Optional argument
+        if "class_labels" in mask:
+            for k, v in list(mask["class_labels"].items()):
+                if (not isinstance(k, numbers.Number)) or (
+                    not isinstance(v, six.string_types)
+                ):
+                    raise TypeError(
+                        "Class labels must be a dictionary of numbers to string"
+                    )
