+++ wandb/sdk/interface/data_types/object_3d.py
@@ -1,0 +1,150 @@
+class Object3D(BatchableMedia):
+    """
+        Wandb class for 3D point clouds.
+
+        Arguments:
+            data_or_path (numpy array, string, io):
+                Object3D can be initialized from a file or a numpy array.
+
+                The file types supported are obj, gltf, babylon, stl.  You can pass a path to
+                    a file or an io object and a file_type which must be one of `'obj', 'gltf', 'babylon', 'stl'`.
+
+                The shape of the numpy array must be one of either:
+                ```
+                [[x y z],       ...] nx3
+                [x y z c],     ...] nx4 where c is a category with supported range [1, 14]
+                [x y z r g b], ...] nx4 where is rgb is color
+                ```
+
+    """
+
+    SUPPORTED_TYPES = set(["obj", "gltf", "babylon", "stl"])
+
+    def __init__(self, data_or_path, **kwargs):
+        super(Object3D, self).__init__()
+
+        if hasattr(data_or_path, "name"):
+            # if the file has a path, we just detect the type and copy it from there
+            data_or_path = data_or_path.name
+
+        if hasattr(data_or_path, "read"):
+            if hasattr(data_or_path, "seek"):
+                data_or_path.seek(0)
+            object3D = data_or_path.read()
+
+            extension = kwargs.pop("file_type", None)
+            if extension == None:
+                raise ValueError(
+                    "Must pass file type keyword argument when using io objects."
+                )
+            if extension not in Object3D.SUPPORTED_TYPES:
+                raise ValueError(
+                    "Object 3D only supports numpy arrays or files of the type: "
+                    + ", ".join(Object3D.SUPPORTED_TYPES)
+                )
+
+            tmp_path = os.path.join(
+                MEDIA_TMP.name, util.generate_id() + "." + extension
+            )
+            with open(tmp_path, "w") as f:
+                f.write(object3D)
+
+            self._set_file(tmp_path, is_tmp=True)
+        elif isinstance(data_or_path, six.string_types):
+            path = data_or_path
+            try:
+                extension = os.path.splitext(data_or_path)[1][1:]
+            except:
+                raise ValueError("File type must have an extension")
+            if extension not in Object3D.SUPPORTED_TYPES:
+                raise ValueError(
+                    "Object 3D only supports numpy arrays or files of the type: "
+                    + ", ".join(Object3D.SUPPORTED_TYPES)
+                )
+
+            self._set_file(data_or_path, is_tmp=False)
+        # Supported different types and scene for 3D scenes
+        elif isinstance(data_or_path, dict) and "type" in data_or_path:
+            if data_or_path["type"] == "lidar/beta":
+                data = {
+                    "type": data_or_path["type"],
+                    "vectors": data_or_path["vectors"].tolist()
+                    if "vectors" in data_or_path
+                    else [],
+                    "points": data_or_path["points"].tolist()
+                    if "points" in data_or_path
+                    else [],
+                    "boxes": data_or_path["boxes"].tolist()
+                    if "boxes" in data_or_path
+                    else [],
+                }
+            else:
+                raise ValueError(
+                    "Type not supported, only 'lidar/beta' is currently supported"
+                )
+
+            tmp_path = os.path.join(MEDIA_TMP.name, util.generate_id() + ".pts.json")
+            json.dump(
+                data,
+                codecs.open(tmp_path, "w", encoding="utf-8"),
+                separators=(",", ":"),
+                sort_keys=True,
+                indent=4,
+            )
+            self._set_file(tmp_path, is_tmp=True, extension=".pts.json")
+        elif is_numpy_array(data_or_path):
+            data = data_or_path
+
+            if len(data.shape) != 2 or data.shape[1] not in {3, 4, 6}:
+                raise ValueError(
+                    """The shape of the numpy array must be one of either
+                                    [[x y z],       ...] nx3
+                                     [x y z c],     ...] nx4 where c is a category with supported range [1, 14]
+                                     [x y z r g b], ...] nx4 where is rgb is color"""
+                )
+
+            data = data.tolist()
+            tmp_path = os.path.join(MEDIA_TMP.name, util.generate_id() + ".pts.json")
+            json.dump(
+                data,
+                codecs.open(tmp_path, "w", encoding="utf-8"),
+                separators=(",", ":"),
+                sort_keys=True,
+                indent=4,
+            )
+            self._set_file(tmp_path, is_tmp=True, extension=".pts.json")
+        else:
+            raise ValueError("data must be a numpy array, dict or a file object")
+
+    @classmethod
+    def get_media_subdir(self):
+        return os.path.join("media", "object3D")
+
+    def to_json(self, run):
+        json_dict = super(Object3D, self).to_json(run)
+        json_dict["_type"] = "object3D-file"
+        return json_dict
+
+    @classmethod
+    def seq_to_json(cls, threeD_list, run, key, step):
+        threeD_list = list(threeD_list)
+
+        jsons = [obj.to_json(run) for obj in threeD_list]
+
+        for obj in jsons:
+            expected = util.to_forward_slash_path(cls.get_media_subdir())
+            if not obj["path"].startswith(expected):
+                raise ValueError(
+                    "Files in an array of Object3D's must be in the {} directory, not {}".format(
+                        expected, obj["path"]
+                    )
+                )
+
+        return {
+            "_type": "object3D",
+            "filenames": [
+                os.path.relpath(j["path"], cls.get_media_subdir()) for j in jsons
+            ],
+            "count": len(jsons),
+            "objects": jsons,
+        }
